int speaker indeces

+++ hw_ss/collate_fn/collate.py
@@ -18,7 +18,7 @@ def collate_fn(dataset_items: List[dict]):
     result_batch["ref_len"] = list(x["ref_len"] for x in dataset_items)
     result_batch["target_len"] = list(x["target_len"] for x in dataset_items)
     result_batch["speaker_id"] = torch.Tensor(
-        list(x["target_id"] for x in dataset_items))
+        list(x["target_id"] for x in dataset_items)).int()
 
     batch_audio_len = min(result_batch["audio_len"])
     batch_target_len = min(result_batch["target_len"])
@@ -51,7 +51,7 @@ def collate_fn_test(dataset_items: List[dict]):
     result_batch["ref_len"] = list(x["ref_len"] for x in dataset_items)
     result_batch["target_len"] = list(x["target_len"] for x in dataset_items)
     result_batch["speaker_id"] = torch.Tensor(
-        list(x["target_id"] for x in dataset_items))
+        list(x["target_id"] for x in dataset_items)).int()
     batch_audio_len = max(result_batch["audio_len"])
     batch_target_len = max(result_batch["target_len"])
     batch_ref_len = max(result_batch["ref_len"])
